+++ src/data/index.ts
@@ -1,2 +1,3 @@
 export * from "./foods";
 export * from "./quirks";
+export * from "./zones";