+++ src/events/index.ts
@@ -5,3 +5,4 @@ export * from "./cooldowns";
 export * from "./select";
 export * from "./resolve";
 export * from "./resolve-social";
+export * from "./reward";